feat(go): dumb golangci config
Make default very dumb so it passes on all but the worst repositories.
I'm working on several repos that have it configured but don't use a config file
and the default settings cause the linters to fail. Working to fix this but
first want a clean ci before ramping up the linters.